third_party/WebKit/LayoutTests/external/wpt/XMLHttpRequest/overridemimetype-blob-expected.txt

This is a testharness.js-based test.
FAIL Use text/xml as fallback MIME type assert_equals: expected "text/xml" but got ""
PASS Use text/xml as fallback MIME type, 2 
FAIL Bogus MIME type should end up as application/octet-stream assert_equals: expected "application/octet-stream" but got ""
FAIL Bogus MIME type should end up as application/octet-stream, 2 assert_equals: expected "application/octet-stream" but got ""
FAIL Valid MIME types need to be normalized assert_equals: expected "hi/x" but got ""
Harness: the test ran to completion.
